handle side effects
