+++ qualia/frontend/src/components/routes/ProtectedRoute.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import { ROUTES } from '@/config/routes';
-import { safeGetLocalStorage } from '@/lib/axios';
+import { safeGetLocalStorage } from '@/lib/storage';
 
 interface ProtectedRouteProps {
   children: ReactNode;
