+++ qualia/frontend/src/components/routes/PublicRoute.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react';
 import { Navigate } from 'react-router-dom';
 import { ROUTES } from '@/config/routes';
-import { safeGetLocalStorage } from '@/lib/axios';
+import { safeGetLocalStorage } from '@/lib/storage';
 
 interface PublicRouteProps {
   children: ReactNode;
